New version with some bugs
