+++ src/data.js
@@ -1,14 +1,14 @@
-const initialCards = [
+export const initialCards = [
 
 ];
 
-const errorMessages = {
+export const errorMessages = {
   empty: 'Это обязательное поле',
   wrongLength: 'Должно быть от 2 до 30 символов',
   wrongUrl: 'Здесь должна быть ссылка'
 };
 
-const config = {
+export const config = {
   url: 'https://praktikum.tk/cohort11',
   authorization: '95676b56-2da6-4da6-b83d-5dd17042dba0',
 }
